Increase processing/polling ratio
My naive health checks report failure after a poll, until the result of that poll
is processed. At the moment, the health checks are very often in a state
of failure, due to this. I should improve the health checks, but for the
moment, I'll just poll less frequently and process more frequently.